backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pydantic==2.11.3
python-dotenv==1.1.0
torch==2.7.1
scikit-learn==1.7.0
xgboost==3.0.2
shap==0.47.2
pandas==3.0.2
numpy==2.4.4
joblib==1.5.1
